+++ CHANGELOG.md
@@ -10,6 +10,10 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 * [mse308072: Clicking on tags in global search results is broken](https://meta.stackexchange.com/q/308072)
 * [mse307976: Can't see all the options while reviewing with a small screen](https://meta.stackexchange.com/q/307976)
 
+**Removed fixes:**
+
+* [mse223737: Inbox heading should be a link](https://meta.stackexchange.com/q/223737)
+
 
 1.52 (2018-03-10)
 ====